+++ lib/cisco_node_utils/upgrade.rb
@@ -70,6 +70,11 @@ module Cisco
       end
     end
 
+    # Return the nxos image installed on the device
+    def self.package
+      config_get('show_version', 'system_image')
+    end
+
     # Return true if box is online and config mode is ready to be used
     def self.box_online?
       output = config_set('upgrade', 'is_box_online')
@@ -93,32 +98,23 @@ module Cisco
     end
 
     # Attempts to upgrade the device to 'image'
-    def self.upgrade(version, image, uri='bootflash:', del_boot=false,
+    def self.upgrade(image, uri='bootflash:', del_boot=false,
                      force_all=false)
-      # Only 'bootflash:' is a supported URI. Fail otherwise.
-      fail "The Uri #{uri} is not supported" unless uri == 'bootflash:'
-      # IMPORTANT - Check if version of image equals the version provided.
-      # This is to avoid entering a loop with the Programmability Agent
-      # continuously trying to reload the device if versions don't match.
-      image_ver = image_version(image, uri)
-      err_str = "Version Mismatch.\n
-                 The version of the image:#{image_ver}\n
-                 The version provided:#{version}\n
-                 Aborting upgrade."
-      fail err_str unless image_ver.to_s.strip == version.to_s.strip
       delete_boot(uri) if del_boot
       force_all ? upgrade_str = 'upgrade_force' : upgrade_str = 'upgrade'
       begin
         Cisco::Logger.debug("Upgrading to version: #{image}")
         config_set('upgrade', upgrade_str, image: image, uri: uri)
-      rescue Cisco::RequestFailed
+      rescue Cisco::RequestFailed, Cisco::CliError => e1
+        # raise if install command failed
+        raise e1 if e1.class == Cisco::CliError
         # Catch 'Backend Processing Error'. Install continues inspite of the
         # error thrown. Resend install command and expect a CliError.
         begin
           config_set('upgrade', upgrade_str, image: image, uri: uri)
-        rescue Cisco::CliError => e
-          raise e unless
-            e.message.include?('Another install procedure may be in progress')
+        rescue Cisco::CliError => e2
+          raise e2 unless
+            e2.message.include?('Another install procedure may be in progress')
         end
       end
     end